+++ src/franz/openrdf/rio/rdfformat.py
@@ -6,11 +6,11 @@
 ################################################################################
 
 from __future__ import unicode_literals
-from builtins import object
-import os
+
+from franz.openrdf.rio.formats import Format
 
 
-class RDFFormat(object):
+class RDFFormat(Format):
     """
     Represents the concept of an RDF data serialization format. RDF formats are
     identified by a {@link #getName() name} and can have one or more associated
@@ -21,31 +21,9 @@ class RDFFormat(object):
     supports_attributes flag is True are capable of encoding triple attributes.
     """
     # A global dictionary mapping extensions to formats
+    # Used by Format.format_for_file_name
     ext_map = {}
 
-    @staticmethod
-    def register(fmt):
-        """ Register a format object."""
-        for ext in fmt.file_extensions:
-            RDFFormat.ext_map['.' + ext.lower()] = fmt
-
-    @staticmethod
-    def rdf_format_for_file_name(filename):
-        """ 
-        Try to guess appropriate RDF format from a file name.
-        Return a pair (format, compression) where format is
-        an RDF format or None (if no matching format was found)
-        and compression is a supported compression method
-        (currently either None or "gzip").
-        """
-        compression = None
-        root, ext = os.path.splitext(filename)
-        if ext.lower() == ".gz":
-            compression = "gzip"
-            _, ext = os.path.splitext(root)
-        fmt = RDFFormat.ext_map.get(ext.lower())
-        return fmt, compression
-            
     def __init__(self, name, mime_types=None, charset="UTF-8",
                  file_extensions=None, supports_namespaces=False,
                  supports_contexts=False, supports_attributes=False,
@@ -67,20 +45,14 @@ class RDFFormat(object):
         :param supports_attributes: If True the format can represent triple
                                     attributes.
         :param register: If True file extensions will be added to the map
-                         used by :func:`RDFFormat.rdf_format_for_file_name`.
+                         used by :meth:`Format.format_for_file_name`.
         """
-        self.name = name
-        self.mime_types = mime_types
-        self.charset = charset
-        self.file_extensions = file_extensions
+        Format.__init__(self, name, mime_types, charset, file_extensions, register)
         self.supports_namespaces = supports_namespaces
         self.supports_contexts = supports_contexts
         self.supports_attributes = supports_attributes
-        if register:
-            RDFFormat.register(self)
 
-    # A call to create_default_formats at the end of this file
-    # will convert these dicts to RDFFormat instances.
+    # These will be automatically converted to RDFFormat instances
 
     RDFXML = dict(
         name="RDF/XML",
@@ -135,5 +107,6 @@ def create_default_formats():
         value = getattr(RDFFormat, name)
         if name.isupper() and isinstance(value, dict):
             setattr(RDFFormat, name, RDFFormat(**value))
+
 
 create_default_formats()